fix layout component and lift page state from bottom nav
